Add destroyer and common tests (#4)
* Add destroyer and common tests

This patch starts to look for correctness under Common Tests. We add a destroyer function, which removes keys from a merkle trie, and we begin to test against common tests. We are currently trying to make sure we properly define branch definitions, which tend to be the most complicated to get right.

* First pass to get a large number of tests passing

* Start to pass a good number of the Merkle Patricia Trie tests