Remove empty trie API function and replace
This patch removes the `empty_trie/0` function from Trie and adds a new `empty_trie_root_hash/1` function. A number of callers expected to be able to call `empty_trie` to get the canonical empty trie, but since all trie root's must be stored in a database, it's impossible to do this without passing in a database. This was causing disparities in what we expected for empty tries (we expected <<>>, but really, it's <<86, 232, 31, 23, 27, 204, 85, 166, 255, 131, 69, 230, 146, 192, 248, 110, 91, 72, 224, 27, 153, 108, 173, 192, 1, 98, 47, 181, 227, 99, 180, 33>>). This patch fixes this issue.