Add back empty trie
While it's annoying to have an empty trie, and it's very likely that will be abused, it's just so darned useful to have as a default value for anything that requires a root hash as it's initial value (e.g. for a Blockchain.Block's state_root, transactions_root and receipts_root). As such, we'll add this back in with a note and why it's a bad idea to overuse it.